Take cdr of frequency list on pick recursion.
The choose-with-frequencies function chooses a random int between
1 and the sum of all frequencies, and then calls pick which recurses
on the frequencies until the correct generator is found.

This commit fixes a bug where the cdr of the frequency list was not
actually passed in the recursive calls, causing a variety of bugs
ranging from always the selecting the first, to never terminating.